refactor: directory path simplification
디렉터리 경로명 단순화. '../../...' 등 불필요하게 긴 디렉터리 경로를 '@'를 사용하여 간소화 시킴.
'@' = './src'
디렉터리 간소화를 위한 npm package eslint-import-resolver-alias 설치.
alias를 위해 eslintrc.js, webpack.config.js 수정.
src 디렉터리 내부 소스 코드 수정.
