+++ src/App.jsx
@@ -1,16 +1,13 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 
-import TestNav from './layouts/TestNav';
-
-import URL from './constants/constants';
-
-import Start from './pages/Start';
-import Interview from './pages/Interview';
-import Result from './pages/Result';
-
-import './styles/Font.scss';
-import './styles/Reset.scss';
+import URL from '@/constants';
+import TestNav from '@/layouts/TestNav';
+import Start from '@/pages/Start';
+import Interview from '@/pages/Interview';
+import Result from '@/pages/Result';
+import '@/styles/Font.scss';
+import '@/styles/Reset.scss';
 
 import './App.scss';
